+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 - Sass plugin: Allow additional importers [#727].
 - Components: Reorder css code to ensure `@import` rules are first.
 - Updated components: `std`, `decap-cms`, `terser` and some icons.
+- Hot reload: use site's location pathname as websocket basepath
 
 ## [2.5.1] - 2025-01-28
 ### Added